[REF] sale_mrp: extract anglo-saxon tests
