+++ addons/sale_mrp/tests/__init__.py
@@ -7,3 +7,4 @@ from . import test_sale_mrp_lead_time
 from . import test_sale_mrp_procurement
 from . import test_multistep_manufacturing
 from . import test_sale_mrp_report
+from . import test_sale_mrp_anglo_saxon_valuation
