+++ addons/sale_mrp/tests/test_sale_mrp_flow.py
@@ -1632,131 +1632,6 @@ class TestSaleMrpFlow(ValuationReconciliationTestCommon):
         self.assertEqual(len(mos), 1)
         self.assertEqual(mos.state, 'cancel')
 
-    def test_12_sale_mrp_anglo_saxon_variant(self):
-        """Test the price unit of kit with variants"""
-        # Check that the correct bom are selected when computing price_unit for COGS
-
-        self.env.company.currency_id = self.env.ref('base.USD')
-        self.uom_unit = self.UoM.create({
-            'name': 'Test-Unit',
-            'category_id': self.categ_unit.id,
-            'factor': 1,
-            'uom_type': 'bigger',
-            'rounding': 1.0})
-        self.company = self.company_data['company']
-        self.company.anglo_saxon_accounting = True
-        self.partner = self.env['res.partner'].create({'name': 'My Test Partner'})
-        self.category = self.env.ref('product.product_category_1').copy({'name': 'Test category', 'property_valuation': 'real_time', 'property_cost_method': 'fifo'})
-        self.account_receiv = self.env['account.account'].create({'name': 'Receivable', 'code': 'RCV00', 'account_type': 'asset_receivable', 'reconcile': True})
-        account_expense = self.env['account.account'].create({'name': 'Expense', 'code': 'EXP00', 'account_type': 'liability_current', 'reconcile': True})
-        account_output = self.env['account.account'].create({'name': 'Output', 'code': 'OUT00', 'account_type': 'liability_current', 'reconcile': True})
-        account_valuation = self.env['account.account'].create({'name': 'Valuation', 'code': 'STV00', 'account_type': 'asset_receivable', 'reconcile': True})
-        self.partner.property_account_receivable_id = self.account_receiv
-        self.category.property_account_income_categ_id = self.account_receiv
-        self.category.property_account_expense_categ_id = account_expense
-        self.category.property_stock_account_input_categ_id = self.account_receiv
-        self.category.property_stock_account_output_categ_id = account_output
-        self.category.property_stock_valuation_account_id = account_valuation
-        self.category.property_stock_journal = self.env['account.journal'].create({'name': 'Stock journal', 'type': 'sale', 'code': 'STK00'})
-
-        # Create variant attributes
-        self.prod_att_1 = self.env['product.attribute'].create({'name': 'Color'})
-        self.prod_attr1_v1 = self.env['product.attribute.value'].create({'name': 'red', 'attribute_id': self.prod_att_1.id, 'sequence': 1})
-        self.prod_attr1_v2 = self.env['product.attribute.value'].create({'name': 'blue', 'attribute_id': self.prod_att_1.id, 'sequence': 2})
-
-        # Create Product template with variants
-        self.product_template = self.env['product.template'].create({
-            'name': 'Product Template',
-            'type': 'product',
-            'uom_id': self.uom_unit.id,
-            'invoice_policy': 'delivery',
-            'categ_id': self.category.id,
-            'attribute_line_ids': [(0, 0, {
-                'attribute_id': self.prod_att_1.id,
-                'value_ids': [(6, 0, [self.prod_attr1_v1.id, self.prod_attr1_v2.id])]
-            })]
-        })
-
-        # Get product variant
-        self.pt_attr1_v1 = self.product_template.attribute_line_ids[0].product_template_value_ids[0]
-        self.pt_attr1_v2 = self.product_template.attribute_line_ids[0].product_template_value_ids[1]
-        self.variant_1 = self.product_template._get_variant_for_combination(self.pt_attr1_v1)
-        self.variant_2 = self.product_template._get_variant_for_combination(self.pt_attr1_v2)
-
-        def create_simple_bom_for_product(product, name, price):
-            component = self.env['product.product'].create({
-                'name': 'Component ' + name,
-                'type': 'product',
-                'uom_id': self.uom_unit.id,
-                'categ_id': self.category.id,
-                'standard_price': price
-            })
-            self.env['stock.quant'].sudo().create({
-                'product_id': component.id,
-                'location_id': self.company_data['default_warehouse'].lot_stock_id.id,
-                'quantity': 10.0,
-            })
-            bom = self.env['mrp.bom'].create({
-                'product_tmpl_id': self.product_template.id,
-                'product_id': product.id,
-                'product_qty': 1.0,
-                'type': 'phantom'
-            })
-            self.env['mrp.bom.line'].create({
-                'product_id': component.id,
-                'product_qty': 1.0,
-                'bom_id': bom.id
-            })
-
-        create_simple_bom_for_product(self.variant_1, "V1", 20)
-        create_simple_bom_for_product(self.variant_2, "V2", 10)
-
-        def create_post_sale_order(product):
-            so_vals = {
-                'partner_id': self.partner.id,
-                'partner_invoice_id': self.partner.id,
-                'partner_shipping_id': self.partner.id,
-                'order_line': [(0, 0, {
-                    'name': product.name,
-                    'product_id': product.id,
-                    'product_uom_qty': 2,
-                    'product_uom': product.uom_id.id,
-                    'price_unit': product.list_price
-                })],
-                'pricelist_id': self.env.ref('product.list0').id,
-                'company_id': self.company.id,
-            }
-            so = self.env['sale.order'].create(so_vals)
-            # Validate the SO
-            so.action_confirm()
-            # Deliver the three finished products
-            pick = so.picking_ids
-            # To check the products on the picking
-            wiz_act = pick.button_validate()
-            wiz = Form(self.env[wiz_act['res_model']].with_context(wiz_act['context'])).save()
-            wiz.process()
-            # Create the invoice
-            so._create_invoices()
-            invoice = so.invoice_ids
-            invoice.action_post()
-            return invoice
-
-        # Create a SO for variant 1
-        self.invoice_1 = create_post_sale_order(self.variant_1)
-        self.invoice_2 = create_post_sale_order(self.variant_2)
-
-        def check_cogs_entry_values(invoice, expected_value):
-            aml = invoice.line_ids
-            aml_expense = aml.filtered(lambda l: l.is_anglo_saxon_line and l.debit > 0)
-            aml_output = aml.filtered(lambda l: l.is_anglo_saxon_line and l.credit > 0)
-            self.assertEqual(aml_expense.debit, expected_value, "Cost of Good Sold entry missing or mismatching for variant")
-            self.assertEqual(aml_output.credit, expected_value, "Cost of Good Sold entry missing or mismatching for variant")
-
-        # Check that the cost of Good Sold entries for variant 1 are equal to 2 * 20 = 40
-        check_cogs_entry_values(self.invoice_1, 40)
-        # Check that the cost of Good Sold entries for variant 2 are equal to 2 * 10 = 20
-        check_cogs_entry_values(self.invoice_2, 20)
-
     def test_13_so_return_kit(self):
         """
         Test that when returning a SO containing only a kit that contains another kit, the
@@ -1927,120 +1802,6 @@ class TestSaleMrpFlow(ValuationReconciliationTestCommon):
         so.action_draft()
         so.action_confirm()
         self.assertEqual(len(so.picking_ids), 1, "The product was already delivered, no need to re-create a delivery order")
-
-    def test_anglo_saxo_return_and_credit_note(self):
-        """
-        When posting a credit note for a returned kit, the value of the anglo-saxo lines
-        should be based on the returned component's value
-        """
-        stock_input_account, stock_output_account, stock_valuation_account, expense_account, stock_journal = _create_accounting_data(self.env)
-        fifo = self.env['product.category'].create({
-            'name': 'FIFO',
-            'property_valuation': 'real_time',
-            'property_cost_method': 'fifo',
-            'property_stock_account_input_categ_id': stock_input_account.id,
-            'property_stock_account_output_categ_id': stock_output_account.id,
-            'property_stock_valuation_account_id': stock_valuation_account.id,
-            'property_stock_journal': stock_journal.id,
-        })
-
-        kit = self._create_product('Simple Kit', self.uom_unit)
-        (kit + self.component_a).categ_id = fifo
-        kit.property_account_expense_id = expense_account
-
-        self.env['mrp.bom'].create({
-            'product_tmpl_id': kit.product_tmpl_id.id,
-            'product_qty': 1.0,
-            'type': 'phantom',
-            'bom_line_ids': [(0, 0, {'product_id': self.component_a.id, 'product_qty': 1.0})]
-        })
-
-        # Receive 3 components: one @10, one @20 and one @60
-        in_moves = self.env['stock.move'].create([{
-            'name': 'IN move @%s' % p,
-            'product_id': self.component_a.id,
-            'location_id': self.env.ref('stock.stock_location_suppliers').id,
-            'location_dest_id': self.company_data['default_warehouse'].lot_stock_id.id,
-            'product_uom': self.component_a.uom_id.id,
-            'product_uom_qty': 1,
-            'price_unit': p,
-        } for p in [10, 20, 60]])
-        in_moves._action_confirm()
-        in_moves.quantity_done = 1
-        in_moves._action_done()
-
-        # Sell 3 kits
-        so = self.env['sale.order'].create({
-            'partner_id': self.env.ref('base.res_partner_1').id,
-            'order_line': [
-                (0, 0, {
-                    'name': kit.name,
-                    'product_id': kit.id,
-                    'product_uom_qty': 3.0,
-                    'product_uom': kit.uom_id.id,
-                    'price_unit': 100,
-                    'tax_id': False,
-                })],
-        })
-        so.action_confirm()
-
-        # Deliver the components: 1@10, then 1@20 and then 1@60
-        pickings = []
-        picking = so.picking_ids
-        while picking:
-            pickings.append(picking)
-            picking.move_ids.quantity_done = 1
-            action = picking.button_validate()
-            if isinstance(action, dict):
-                wizard = Form(self.env[action['res_model']].with_context(action['context'])).save()
-                wizard.process()
-            picking = picking.backorder_ids
-
-        invoice = so._create_invoices()
-        invoice.action_post()
-
-        # Receive one @100
-        in_moves = self.env['stock.move'].create({
-            'name': 'IN move @100',
-            'product_id': self.component_a.id,
-            'location_id': self.env.ref('stock.stock_location_suppliers').id,
-            'location_dest_id': self.company_data['default_warehouse'].lot_stock_id.id,
-            'product_uom': self.component_a.uom_id.id,
-            'product_uom_qty': 1,
-            'price_unit': 100,
-        })
-        in_moves._action_confirm()
-        in_moves.quantity_done = 1
-        in_moves._action_done()
-
-        # Return the second picking (i.e. one component @20)
-        ctx = {'active_id': pickings[1].id, 'active_model': 'stock.picking'}
-        return_wizard = Form(self.env['stock.return.picking'].with_context(ctx)).save()
-        return_picking_id, dummy = return_wizard._create_returns()
-        return_picking = self.env['stock.picking'].browse(return_picking_id)
-        return_picking.move_ids.quantity_done = 1
-        return_picking.button_validate()
-
-        # Add a credit note for the returned kit
-        ctx = {'active_model': 'account.move', 'active_ids': invoice.ids}
-        refund_wizard = self.env['account.move.reversal'].with_context(ctx).create({
-            'refund_method': 'refund',
-            'journal_id': invoice.journal_id.id,
-        })
-        action = refund_wizard.reverse_moves()
-        reverse_invoice = self.env['account.move'].browse(action['res_id'])
-        with Form(reverse_invoice) as reverse_invoice_form:
-            with reverse_invoice_form.invoice_line_ids.edit(0) as line:
-                line.quantity = 1
-        reverse_invoice.action_post()
-
-        amls = reverse_invoice.line_ids
-        stock_out_aml = amls.filtered(lambda aml: aml.account_id == stock_output_account)
-        self.assertEqual(stock_out_aml.debit, 20, 'Should be to the value of the returned component')
-        self.assertEqual(stock_out_aml.credit, 0)
-        cogs_aml = amls.filtered(lambda aml: aml.account_id == expense_account)
-        self.assertEqual(cogs_aml.debit, 0)
-        self.assertEqual(cogs_aml.credit, 20, 'Should be to the value of the returned component')
 
     def test_kit_margin_and_return_picking(self):
         """ This test ensure that, when returning the components of a sold kit, the
